coc: Remove broken coc-yaml.
Unfortunately the schema checking for GitHub is outdated, and despite
apparently having a way to pull the latest schema from SchemaStore (and
that setting being enabled by default), there doesn't seem to be any
obvious way to actually update the schema. This plugin also seems to be
abandoned for over a year.